Improve dummy authorization handlers
* Move them outside of decidim.
* Install them optional via the `--demo` option.
* Merge the "example" one with the "dummy" one. Their purpose was the
  same: explain the feature, so we merge both taking the best of each
  (docs + real working example).
* Add another dummy handler, so the "handler" selection features can be
  used on development applications by default.
