+++ decidim-core/config/locales/en.yml
@@ -110,6 +110,10 @@ en:
       photos:
         related_photos: Related photos
     authorization_handlers:
+      another_dummy_authorization_handler:
+        explanation: Get verified by introducing a passport number starting with "A"
+        fields:
+          passport_number: Passport number
       dummy_authorization_handler:
         explanation: Get verified by introducing a document number ending with "X"
         fields:
